mini_calling_service/pyproject.toml
[project]
name = "mini-calling-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "livekit-agents[groq,silero,turn-detector]==1.0.22",
    "livekit-plugins-elevenlabs==1.0.20",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]
